+++ src/main/java/com/bandhanbook/app/payload/request/UserRegisterRequest.java
@@ -24,7 +24,7 @@ public class UserRegisterRequest {
     @NotNull(message = "Gender must not be null")
     private GenderOptions gender;
 
-    private int registrationFee;
+    private double registrationFee;
 
     private Date dob;
