+++ src/main/java/com/bandhanbook/app/payload/response/EventResponse.java
@@ -21,7 +21,7 @@ public class EventResponse {
     private String organizationId;
     private String location;
     private Status status;
-    private int registrationFee;
+    private double registrationFee;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private LocalDateTime createdAt;
